worker/requirements.txt
celery[redis]
psycopg2-binary
numpy
pandas
tensorflow
scikit-learn
geojson
simplekml
